Refactor quiz configuration cache into dedicated module
